Pagination Broken needs to be fixed

+++ inventory/migrations/2022-03-09-134654_add_rank_to_products/up.sql
@@ -1,2 +1,2 @@
 -- Your SQL goes here
-ALTER TABLE products ADD COLUMN product_rank NUMERIC DEFAULT 0.0;
+ALTER TABLE products ADD COLUMN product_rank FLOAT DEFAULT 0.0;
